+++ src/test/java/com/nexmo/client/account/CreateSecretMethodTest.java
@@ -76,7 +76,7 @@ public class CreateSecretMethodTest {
 
     @Test
     public void testCustomUri() throws Exception {
-        HttpWrapper wrapper = new HttpWrapper(new HttpConfig.Builder().baseUri("https://example.com").build());
+        HttpWrapper wrapper = new HttpWrapper(HttpConfig.builder().baseUri("https://example.com").build());
         CreateSecretMethod method = new CreateSecretMethod(wrapper);
         CreateSecretRequest request = new CreateSecretRequest("account-id", "secret");