+++ src/test/java/com/nexmo/client/account/PrefixPricingMethodTest.java
@@ -51,7 +51,7 @@ public class PrefixPricingMethodTest {
 
     @Test
     public void testCustomUri() throws Exception {
-        HttpWrapper wrapper = new HttpWrapper(new HttpConfig.Builder().baseUri("https://example.com").build());
+        HttpWrapper wrapper = new HttpWrapper(HttpConfig.builder().baseUri("https://example.com").build());
         PrefixPricingMethod method = new PrefixPricingMethod(wrapper);
         PrefixPricingRequest request = new PrefixPricingRequest(ServiceType.VOICE, "prefix");