+++ src/test/java/com/nexmo/client/applications/CreateApplicationMethodTest.java
@@ -149,7 +149,7 @@ public class CreateApplicationMethodTest {
 
     @Test
     public void testCustomUri() throws Exception {
-        HttpWrapper wrapper = new HttpWrapper(new HttpConfig.Builder().baseUri("https://example.com").build());
+        HttpWrapper wrapper = new HttpWrapper(HttpConfig.builder().baseUri("https://example.com").build());
         CreateApplicationMethod method = new CreateApplicationMethod(wrapper);
         CreateApplicationRequest request = new CreateApplicationRequest("name", "answer-url", "event-url");