+++ src/test/java/com/nexmo/client/applications/GetApplicationEndpointTest.java
@@ -107,7 +107,7 @@ public class GetApplicationEndpointTest {
 
     @Test
     public void testCustomUri() throws Exception {
-        HttpWrapper wrapper = new HttpWrapper(new HttpConfig.Builder().baseUri("https://example.com").build());
+        HttpWrapper wrapper = new HttpWrapper(HttpConfig.builder().baseUri("https://example.com").build());
         GetApplicationEndpoint endpoint = new GetApplicationEndpoint(wrapper);
 
         RequestBuilder builder = endpoint.makeRequest("application-id");