+++ src/test/java/com/nexmo/client/applications/ListApplicationsEndpointTest.java
@@ -148,7 +148,7 @@ public class ListApplicationsEndpointTest {
 
     @Test
     public void testCustomUri() throws Exception {
-        HttpWrapper wrapper = new HttpWrapper(new HttpConfig.Builder().baseUri("https://example.com").build());
+        HttpWrapper wrapper = new HttpWrapper(HttpConfig.builder().baseUri("https://example.com").build());
         ListApplicationsEndpoint endpoint = new ListApplicationsEndpoint(wrapper);
         ListApplicationsRequest request = new ListApplicationsRequest();
 
@@ -175,7 +175,7 @@ public class ListApplicationsEndpointTest {
 
     @Test
     public void testCustomUriWithPage() throws Exception {
-        HttpWrapper wrapper = new HttpWrapper(new HttpConfig.Builder().baseUri("https://example.com").build());
+        HttpWrapper wrapper = new HttpWrapper(HttpConfig.builder().baseUri("https://example.com").build());
         ListApplicationsEndpoint endpoint = new ListApplicationsEndpoint(wrapper);
         ListApplicationsRequest request = new ListApplicationsRequest();
         request.setPageIndex(32);