+++ src/test/java/com/nexmo/client/numbers/SearchNumbersEndpointTest.java
@@ -136,7 +136,7 @@ public class SearchNumbersEndpointTest {
 
     @Test
     public void testCustomUri() throws Exception {
-        HttpWrapper wrapper = new HttpWrapper(new HttpConfig.Builder().baseUri("https://example.com").build());
+        HttpWrapper wrapper = new HttpWrapper(HttpConfig.builder().baseUri("https://example.com").build());
         SearchNumbersEndpoint endpoint = new SearchNumbersEndpoint(wrapper);
         SearchNumbersFilter filter = new SearchNumbersFilter("BB");
         filter.setIndex(10);