+++ src/test/java/com/nexmo/client/numbers/UpdateNumberEndpointTest.java
@@ -122,7 +122,7 @@ public class UpdateNumberEndpointTest {
 
     @Test
     public void testCustomUri() throws Exception {
-        HttpWrapper wrapper = new HttpWrapper(new HttpConfig.Builder().baseUri("https://example.com").build());
+        HttpWrapper wrapper = new HttpWrapper(HttpConfig.builder().baseUri("https://example.com").build());
         UpdateNumberEndpoint endpoint = new UpdateNumberEndpoint(wrapper);
         UpdateNumberRequest request = new UpdateNumberRequest("447700900013", "UK");