+++ src/test/java/com/nexmo/client/sms/SearchSmsEndpointTest.java
@@ -151,7 +151,7 @@ public class SearchSmsEndpointTest {
 
     @Test
     public void testCustomUri() throws Exception {
-        HttpWrapper wrapper = new HttpWrapper(new HttpConfig.Builder().baseUri("https://example.com").build());
+        HttpWrapper wrapper = new HttpWrapper(HttpConfig.builder().baseUri("https://example.com").build());
         SmsSearchEndpoint endpoint = new SmsSearchEndpoint(wrapper);
         SmsIdSearchRequest request = new SmsIdSearchRequest("one-id");