+++ src/test/java/com/nexmo/client/sns/SnsEndpointTest.java
@@ -158,7 +158,7 @@ public class SnsEndpointTest {
 
     @Test
     public void testCustomUri() throws Exception {
-        HttpWrapper wrapper = new HttpWrapper(new HttpConfig.Builder().baseUri("https://example.com").build());
+        HttpWrapper wrapper = new HttpWrapper(HttpConfig.builder().baseUri("https://example.com").build());
         SnsEndpoint endpoint = new SnsEndpoint(wrapper);
         SnsRequest request = new SnsPublishRequest("to", "arn", "from", "message");