+++ src/test/java/com/nexmo/client/verify/ControlEndpointTest.java
@@ -100,7 +100,7 @@ public class ControlEndpointTest {
 
     @Test
     public void testCustomUri() throws Exception {
-        HttpWrapper wrapper = new HttpWrapper(new HttpConfig.Builder().baseUri("https://example.com").build());
+        HttpWrapper wrapper = new HttpWrapper(HttpConfig.builder().baseUri("https://example.com").build());
         ControlEndpoint endpoint = new ControlEndpoint(wrapper);
         ControlRequest request = new ControlRequest("request-id", VerifyControlCommand.CANCEL);