+++ src/test/java/com/nexmo/client/verify/SearchMethodTest.java
@@ -48,7 +48,7 @@ public class SearchMethodTest {
 
     @Test
     public void testCustomUri() throws Exception {
-        HttpWrapper wrapper = new HttpWrapper(new HttpConfig.Builder().baseUri("https://example.com").build());
+        HttpWrapper wrapper = new HttpWrapper(HttpConfig.builder().baseUri("https://example.com").build());
         SearchMethod method = new SearchMethod(wrapper);
         SearchRequest request = new SearchRequest("request-id");