+++ src/test/java/com/nexmo/client/verify/VerifyMethodTest.java
@@ -122,7 +122,7 @@ public class VerifyMethodTest extends MethodTest<VerifyMethod> {
 
     @Test
     public void testCustomUri() throws Exception {
-        HttpWrapper wrapper = new HttpWrapper(new HttpConfig.Builder().baseUri("https://example.com").build());
+        HttpWrapper wrapper = new HttpWrapper(HttpConfig.builder().baseUri("https://example.com").build());
         VerifyMethod method = new VerifyMethod(wrapper);
         VerifyRequest request = new VerifyRequest("4477990090090",
                 "Brand.com",